Make gas type be u128 (#319)
* free ci space to avoid errors

* dont free on macos

* needs sudo

* needs sudo

* make gas u128

* fix bench

* modern way

* remove comment

+++ examples/starknet.rs
@@ -20,7 +20,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
     fn get_block_hash(
         &mut self,
         block_number: u64,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<cairo_felt::Felt252> {
         println!("Called `get_block_hash({block_number})` from MLIR.");
         Ok(Felt252::from_bytes_be(b"get_block_hash ok"))
@@ -28,7 +28,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
 
     fn get_execution_info(
         &self,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<cairo_native::starknet::ExecutionInfo> {
         println!("Called `get_execution_info()` from MLIR.");
         Ok(ExecutionInfo {
@@ -58,7 +58,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         contract_address_salt: cairo_felt::Felt252,
         calldata: &[cairo_felt::Felt252],
         deploy_from_zero: bool,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<(cairo_felt::Felt252, Vec<cairo_felt::Felt252>)> {
         println!("Called `deploy({class_hash}, {contract_address_salt}, {calldata:?}, {deploy_from_zero})` from MLIR.");
         Ok((
@@ -70,7 +70,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
     fn replace_class(
         &mut self,
         class_hash: cairo_felt::Felt252,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<()> {
         println!("Called `replace_class({class_hash})` from MLIR.");
         Ok(())
@@ -81,7 +81,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         class_hash: cairo_felt::Felt252,
         function_selector: cairo_felt::Felt252,
         calldata: &[cairo_felt::Felt252],
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Vec<cairo_felt::Felt252>> {
         println!(
             "Called `library_call({class_hash}, {function_selector}, {calldata:?})` from MLIR."
@@ -94,7 +94,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         address: cairo_felt::Felt252,
         entry_point_selector: cairo_felt::Felt252,
         calldata: &[cairo_felt::Felt252],
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Vec<cairo_felt::Felt252>> {
         println!(
             "Called `call_contract({address}, {entry_point_selector}, {calldata:?})` from MLIR."
@@ -106,7 +106,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &mut self,
         address_domain: u32,
         address: cairo_felt::Felt252,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<cairo_felt::Felt252> {
         println!("Called `storage_read({address_domain}, {address})` from MLIR.");
         Ok(address * &Felt252::new(3))
@@ -117,7 +117,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         address_domain: u32,
         address: cairo_felt::Felt252,
         value: cairo_felt::Felt252,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<()> {
         println!("Called `storage_write({address_domain}, {address}, {value})` from MLIR.");
         Ok(())
@@ -127,7 +127,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &mut self,
         keys: &[cairo_felt::Felt252],
         data: &[cairo_felt::Felt252],
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<()> {
         println!("Called `emit_event({keys:?}, {data:?})` from MLIR.");
         Ok(())
@@ -137,13 +137,17 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &mut self,
         to_address: cairo_felt::Felt252,
         payload: &[cairo_felt::Felt252],
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<()> {
         println!("Called `send_message_to_l1({to_address}, {payload:?})` from MLIR.");
         Ok(())
     }
 
-    fn keccak(&self, input: &[u64], _gas: &mut u64) -> SyscallResult<cairo_native::starknet::U256> {
+    fn keccak(
+        &self,
+        input: &[u64],
+        _gas: &mut u128,
+    ) -> SyscallResult<cairo_native::starknet::U256> {
         println!("Called `keccak({input:?})` from MLIR.");
         Ok(U256(Felt252::from(1234567890).to_le_bytes()))
     }
@@ -152,7 +156,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &mut self,
         _p0: cairo_native::starknet::Secp256k1Point,
         _p1: cairo_native::starknet::Secp256k1Point,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -161,7 +165,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _x: cairo_native::starknet::U256,
         _y_parity: bool,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -169,7 +173,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
     fn secp256k1_get_xy(
         &self,
         _p: cairo_native::starknet::Secp256k1Point,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<(cairo_native::starknet::U256, cairo_native::starknet::U256)> {
         todo!()
     }
@@ -178,7 +182,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _p: cairo_native::starknet::Secp256k1Point,
         _m: cairo_native::starknet::U256,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -187,7 +191,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _x: cairo_native::starknet::U256,
         _y: cairo_native::starknet::U256,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -196,7 +200,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _p0: cairo_native::starknet::Secp256k1Point,
         _p1: cairo_native::starknet::Secp256k1Point,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -205,7 +209,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _x: cairo_native::starknet::U256,
         _y_parity: bool,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -213,7 +217,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
     fn secp256r1_get_xy(
         &self,
         _p: cairo_native::starknet::Secp256k1Point,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<(cairo_native::starknet::U256, cairo_native::starknet::U256)> {
         todo!()
     }
@@ -222,7 +226,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &self,
         _p: cairo_native::starknet::Secp256k1Point,
         _m: cairo_native::starknet::U256,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
@@ -231,7 +235,7 @@ impl StarkNetSyscallHandler for SyscallHandler {
         &mut self,
         _x: cairo_native::starknet::U256,
         _y: cairo_native::starknet::U256,
-        _gas: &mut u64,
+        _gas: &mut u128,
     ) -> SyscallResult<Option<cairo_native::starknet::Secp256k1Point>> {
         todo!()
     }
